Change : Phone number type from int to long :

+++ Migrations/AppDbContextModelSnapshot.cs
@@ -390,8 +390,8 @@ namespace ComputerStoreWebApi.Migrations
                         .IsRequired()
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("PhoneNumber")
-                        .HasColumnType("int");
+                    b.Property<long>("PhoneNumber")
+                        .HasColumnType("bigint");
 
                     b.HasKey("Id");
